requirements.txt
# Core scientific computing libraries
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# Single-cell analysis
scanpy>=1.8.0
anndata>=0.8.0

# Machine learning and data visualization
scikit-learn>=1.0.0
seaborn>=0.11.0
matplotlib>=3.5.0
umap-learn>=0.5.0

# Pathway analysis
gseapy>=1.0.0

# Clustering analysis
pyclustree>=0.1.0

# Color palettes
colorcet>=2.0.0

# Progress bars
tqdm>=4.62.0

# LLM client
google-generativeai>=0.8.0

# Batch integration (optional)
harmonypy>=0.0.6

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Development and testing
pytest>=7.0.0